[fix] lever 3 : QueryDsl 메소드를 List<Dto> -> List<Entity> 방식으로 변경, BatchSize 적용

+++ src/main/java/org/example/expert/domain/todo/dto/response/TodoSearchResponse.java
@@ -5,10 +5,10 @@ import lombok.Getter;
 @Getter
 public class TodoSearchResponse {
     private final String title;
-    private final Long managerCount;
-    private final Long commentCount;
+    private final int managerCount;
+    private final int commentCount;
 
-    public TodoSearchResponse(String title, Long managerCount, Long commentCount) {
+    public TodoSearchResponse(String title, int managerCount, int commentCount) {
         this.title = title;
         this.managerCount = managerCount;
         this.commentCount = commentCount;
